Move provision script in its own folder
* Segment the file in multiple files (helpers)
* Update the studio flow if one already exists
